+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Change primary Linux kernel from linux-fsync to linux-nobara.
 - Change more first-time setup tasks to be included in the gaming section instead of separate sections.
 - Add Proton Sarek for legacy graphics cards.
+- Add hibernation support for NVIDIA graphics cards.
 
 ## 4.1.0 - The Dual-Boot Update